--- SpellChecker.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30204.135
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SpellChecker", "SpellChecker\SpellChecker.vcxproj", "{36FB577A-CF97-4C69-A863-E5888D38E1FD}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{36FB577A-CF97-4C69-A863-E5888D38E1FD}.Debug|x64.ActiveCfg = Debug|x64
-		{36FB577A-CF97-4C69-A863-E5888D38E1FD}.Debug|x64.Build.0 = Debug|x64
-		{36FB577A-CF97-4C69-A863-E5888D38E1FD}.Debug|x86.ActiveCfg = Debug|Win32
-		{36FB577A-CF97-4C69-A863-E5888D38E1FD}.Debug|x86.Build.0 = Debug|Win32
-		{36FB577A-CF97-4C69-A863-E5888D38E1FD}.Release|x64.ActiveCfg = Release|x64
-		{36FB577A-CF97-4C69-A863-E5888D38E1FD}.Release|x64.Build.0 = Release|x64
-		{36FB577A-CF97-4C69-A863-E5888D38E1FD}.Release|x86.ActiveCfg = Release|Win32
-		{36FB577A-CF97-4C69-A863-E5888D38E1FD}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {71CA97C6-9CC5-4057-9025-A1AA2ED8208C}
-	EndGlobalSection
-EndGlobal
