--- SpellChecker/resources/german.dic
@@ -1,1 +1,0 @@
-﻿
