--- SpellChecker/x64/Release/SpellChecker.tlog/SpellChecker.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v142:VCToolArchitecture=Native32Bit:VCToolsVersion=14.26.28801:TargetPlatformVersion=10.0.18362.0:
-Release|x64|F:\Programmieren\Spellcheck\SpellChecker\|
