+++ resources/german.dic
@@ -1,0 +1,1 @@
+﻿
